Refactor PWM Output generator libary and add thread-safety #10.
