+++ include/definitions.h
@@ -24,7 +24,7 @@
 #define I2C_FREQ 100000
 
 #define I2C_ADDRESS_ADS1x15 0x48
-
+#define I2C_ADDRESS_PCA9685 0x40
 
 /*
  *  SPI
