+++ lib/PWM/PWM.h
@@ -1,0 +1,11 @@
+//
+// 16-Channel Pulse Width Modulation
+//
+
+#ifndef PWM_H
+#define PWM_H
+
+void init_pwm(void);
+void update_pwm_demo_task(void *pvParameter);
+
+#endif // PWM_H
